requirements.txt
numpy
easyocr
opencv-python
ultralytics
python-dotenv
mysql-connector-python